Migrate Rise to separate implementation
